test through fromReflection
