+++ packages/oats-runtime/test/make.spec.ts
@@ -1,5 +1,4 @@
 import * as make from '../src/make';
-import * as jsc from 'jsverify';
 import { ShapeOfTestClass, TestClass } from './test-class';
 import { validationErrorPrinter } from '../src/make';
 
@@ -33,34 +32,6 @@ describe('Make', () => {
 });
 
 describe('makeOneOf', () => {
-  it('succeeds if only one match', () => {
-    const fun = make.makeOneOf(
-      make.makeObject({ a: make.makeString() }),
-      make.makeObject({ a: make.makeNumber() })
-    );
-    expect(fun({ a: 1 }).success()).toEqual({ a: 1 });
-  });
-
-  it('prefers matching ValueClass', () => {
-    const fun = make.makeOneOf(make.makeAny(), TestClass.make);
-    const test = TestClass.make({ a: ['a'], b: 'b' }).success();
-    expect(fun(test).success()).toEqual(test);
-  });
-
-  it('fails if multiple values classes match', () => {
-    const fun = make.makeOneOf(make.makeAny(), TestClass.make, TestClass.make);
-    const test = TestClass.make({ a: ['a'], b: 'b' }).success();
-    expect(fun(test).isSuccess()).toBeFalsy();
-  });
-
-  it('fails on multiple matches', () => {
-    const fun = make.makeOneOf(
-      make.makeObject({ a: make.makeString() }),
-      make.makeObject({ a: make.makeAny() })
-    );
-    expect(fun({ a: 'x' }).isSuccess()).toBeFalsy();
-  });
-
   it('groups errors', () => {
     const fun = make.makeObject({
       root: make.makeOneOf(
@@ -86,184 +57,5 @@ describe('registerFormat', () => {
     expect(() => {
       make.registerFormat('duplicate-register', () => make.Make.ok(undefined));
     }).toThrow('format duplicate-register is already registered');
-  });
-});
-
-describe('makeString', () => {
-  it('rejects if the pattern does not match', () => {
-    const fun = make.makeString(undefined, 'a+');
-    expect(fun('b').errors[0].error).toEqual('b does not match pattern /a+/');
-  });
-
-  it('accepts if the pattern matches', () => {
-    const fun = make.makeString(undefined, 'a+');
-    fun('aaa').success();
-  });
-
-  it('throws on invalid regex', () => {
-    expect(() => make.makeString(undefined, '\\')).toThrow(
-      "pattern for 'type: string' is not valid: Invalid regular expression"
-    );
-  });
-
-  it('accepts if format is not defined', () => {
-    const fun = make.makeString('some-format');
-    fun('b').success();
-  });
-
-  it('rejects if format rejects', () => {
-    make.registerFormat('some-rejecting-format', () =>
-      make.Make.error([{ path: [], error: 'some error' }])
-    );
-    const fun = make.makeString('some-rejecting-format');
-    expect(fun('b').errors[0].error).toEqual('some error');
-  });
-
-  it('accepts if format accepts', () => {
-    make.registerFormat('some-accepting-format', () => make.Make.ok(undefined));
-    const fun = make.makeString('some-accepting-format');
-    expect(fun('b').success()).toEqual('b');
-  });
-});
-
-describe('makeAllOf', () => {
-  it('applies all makers to value in succession', () => {
-    const fun = make.makeAllOf(
-      make.makeObject({ a: make.makeString() }, make.makeAny()),
-      make.makeObject({}, make.makeAllOf())
-    );
-    expect(fun({ a: 'xxx', b: 1 }).success()).toEqual({ a: 'xxx', b: 1 });
-  });
-});
-
-describe('makeAny', () => {
-  jsc.property('allows anything', jsc.json, async item => {
-    const fun = make.makeAny();
-    expect(fun(item).success()).toEqual(item);
-    if (item && typeof item === 'object') {
-      expect(fun(item).success() !== item).toBeTruthy();
-    }
-    return true;
-  });
-});
-
-describe('makeNumber', () => {
-  it('enforces minimum if passed', () => {
-    const fun = make.makeNumber(3);
-    expect(fun(2).errors[0].error).toMatch('expected a number greater or equal to');
-    expect(fun(3).isSuccess()).toBeTruthy();
-  });
-  it('enforces maximum if passed', () => {
-    const fun = make.makeNumber(undefined, 3);
-    expect(fun(4).errors[0].error).toMatch('expected a number smaller or equal to');
-    expect(fun(3).isSuccess()).toBeTruthy();
-  });
-});
-
-describe('makeArray', () => {
-  it('keeps the order', () => {
-    const fun = make.makeArray(make.makeString());
-    expect(fun(['a', 'b']).success()).toEqual(['a', 'b']);
-  });
-  it('enforces min size if passed', () => {
-    const fun = make.makeArray(make.makeString(), 3);
-    expect(fun(['a', 'b']).errors[0].error).toMatch('expected an array of minimum length');
-    expect(fun(['a', 'b', 'c']).isSuccess()).toBeTruthy();
-  });
-  it('enforces max size if passed', () => {
-    const fun = make.makeArray(make.makeString(), undefined, 3);
-    expect(fun(['a', 'b', 'c', 'd']).errors[0].error).toMatch(
-      'expected an array of maximum length'
-    );
-    expect(fun(['a', 'b', 'c']).isSuccess()).toBeTruthy();
-  });
-});
-
-describe('makeObject', () => {
-  describe('magic fields', () => {
-    it('disallows __proto__', () => {
-      const obj = Object.create(null);
-      obj.__proto__ = make.makeNumber();
-      const fun = make.makeObject(obj);
-      const value = JSON.parse(`{"__proto__": 1}`);
-      expect(fun(value).errors[0].error).toEqual(
-        'Using __proto__ as field of an object is not allowed'
-      );
-    });
-
-    it('disallows array', () => {
-      const fun = make.makeObject({}, make.makeAny());
-      expect(fun([]).errors[0].error).toEqual('expected an object, but got "[]" instead.');
-    });
-
-    it('disallows constructor', () => {
-      const obj = Object.create(null);
-      obj.constructor = make.makeNumber();
-      const fun = make.makeObject(obj);
-      const value = JSON.parse(`{"constructor": 1}`);
-      expect(fun(value).errors[0].error).toEqual(
-        'Using constructor as field of an object is not allowed'
-      );
-    });
-
-    it('disallows __proto__ in additionalFields', () => {
-      const fun = make.makeObject({}, make.makeNumber());
-      const value = JSON.parse(`{"__proto__": 1}`);
-      expect(fun(value).errors[0].error).toEqual(
-        'Using __proto__ as objects additional field is not allowed.'
-      );
-    });
-
-    it('disallows constructor in additionalFields', () => {
-      const fun = make.makeObject({}, make.makeNumber());
-      const value = JSON.parse(`{"constructor": 1}`);
-      expect(fun(value).errors[0].error).toEqual(
-        'Using constructor as objects additional field is not allowed.'
-      );
-    });
-  });
-
-  it('drops unknown properties if told to', () => {
-    const fun = make.makeObject({ a: make.makeNumber() });
-    expect(fun({ a: 1, missing: 'a' }, { unknownField: 'drop' }).success()).toEqual({ a: 1 });
-  });
-
-  it('drops unknown properties in nested objects', () => {
-    const fun = make.makeObject({ a: make.makeObject({ b: make.makeNumber() }) });
-    expect(fun({ a: { b: 1, missing: 'value' } }, { unknownField: 'drop' }).success()).toEqual({
-      a: { b: 1 }
-    });
-  });
-
-  it('drops unknown properties in nested additional prop objects', () => {
-    const fun = make.makeObject({}, make.makeObject({ b: make.makeNumber() }));
-    expect(fun({ a: { b: 1, missing: 'value' } }, { unknownField: 'drop' }).success()).toEqual({
-      a: { b: 1 }
-    });
-  });
-
-  it('disallows extra fields', () => {
-    const fun = make.makeObject({ a: make.makeNumber() });
-    expect(fun({ a: 1, missing: 'a' }).isError()).toBeTruthy();
-  });
-
-  it('allows fields', () => {
-    const fun = make.makeObject({ a: make.makeNumber() });
-    expect(fun({ a: 1 }).isSuccess()).toBeTruthy();
-  });
-
-  it('allows additional props', () => {
-    const fun = make.makeObject({}, make.makeString());
-    expect(fun({ a: 'xxx' }).success()).toEqual({ a: 'xxx' });
-  });
-
-  it('prefers specified field to additional props', () => {
-    const fun = make.makeObject({ a: make.makeNumber() }, make.makeString());
-    expect(fun({ a: 1 }).success()).toEqual({ a: 1 });
-  });
-
-  it('allows undefined props', () => {
-    const fun = make.makeObject({ a: make.makeOptional(make.makeNumber()) });
-    expect(fun({ a: undefined }).success()).toEqual({});
   });
 });
